segregated API functions from index.js into files in routes/api folder
also some further refactoring: moved logAuditAction() into services/audit.js. Moved prisma client initialization into lib/prisma.js
